book_manage/requirements.txt
connexion[swagger-ui]==2.14.2
Flask==2.2.5
Werkzeug==2.2.3
python-dotenv==1.0.1
